+++ api/src/config/playwright.js
@@ -2,11 +2,11 @@
  * This module stores the configuration for using Puppeteer to scrape Google Scholar.
  */
 
-const puppeteerConfig = {
+const playwrightConfig = {
   noOfDummyLinks: 4, // first 4 links aren't publications can be ignored
   noOfThreads: 10, // number of concurrent threads
   pageSize: 10, // number of publications to return at once
-  base: 'https://scholar.google.com.sg',
+  gScholarHome: 'https://scholar.google.com.sg',
   baseUrl: 'https://scholar.google.com.sg/citations?hl=en&user=',
   startSuffix: '&cstart=', // concatenate number to start at after
   pageSizeSuffix: '&pagesize=', // concatenate page size after
@@ -22,4 +22,4 @@ const categoryType = {
 
 const categoryTypeEnum = ['CONFERENCE', 'JOURNAL', 'OTHER', 'BOOK'];
 
-module.exports = { puppeteerConfig, categoryType, categoryTypeEnum };
+module.exports = { playwrightConfig, categoryType, categoryTypeEnum };
