+++ api/src/middleware/publications.js
@@ -4,7 +4,7 @@
 
 const { body, validationResult } = require('express-validator');
 const axios = require('axios');
-const { categoryTypeEnum } = require('../config/puppeteer');
+const { categoryTypeEnum } = require('../config/playwright');
 const { fillErrorObject } = require('./error');
